Actually the thing I legit just tried might work
check out me passing "this" through as the parameter; because then it updates on click like I want it to, but I can also just straight ask if it is checked or not, and I can still get at the value. Not sure if it works longrun though. Only implemented it on acrobatics right now